+++ packages/dependency-graph/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.4.0](https://github.com/typescript-tools/typescript-tools/compare/@typescript-tools/dependency-graph@1.3.1...@typescript-tools/dependency-graph@1.4.0) (2021-01-17)
+
+
+### Features
+
+* add lock-internal-packages ([44e2ed1](https://github.com/typescript-tools/typescript-tools/commit/44e2ed1cebba3a02e48ca5c6bd9c67aa21b178ed))
+
+
+
+
+
 ## [1.3.1](https://github.com/typescript-tools/typescript-tools/compare/@typescript-tools/dependency-graph@1.3.0...@typescript-tools/dependency-graph@1.3.1) (2021-01-05)
 
 **Note:** Version bump only for package @typescript-tools/dependency-graph