+++ packages/merge-package-locks/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.2.0](https://github.com/typescript-tools/typescript-tools/compare/@typescript-tools/merge-package-locks@1.1.0...@typescript-tools/merge-package-locks@1.2.0) (2021-01-17)
+
+
+### Features
+
+* add lock-internal-packages ([44e2ed1](https://github.com/typescript-tools/typescript-tools/commit/44e2ed1cebba3a02e48ca5c6bd9c67aa21b178ed))
+
+
+
+
+
 # 1.1.0 (2021-01-05)